globais -> comum
classe globais, pra acompanhar bipes e as funções
