+++ comum.py
@@ -1,0 +1,79 @@
+from pybricks.tools import wait
+from pybricks.parameters import Color
+
+class globais:
+    # cabeça
+    rodas = None
+
+    # braço
+    motor_garra = None
+    motor_vertical = None
+
+    # comum
+    hub = None
+    ble = None
+    name = None
+    nome = None
+
+    @classmethod
+    def init(self, hub, teste, debug, nome, ble=None):
+        self.TESTE = teste
+        self.DEBUG = debug
+
+        self.hub = hub
+        self.ble = ble or hub.ble
+
+        self.name = hub.system.name()
+        self.nome = nome
+        self.globalizar()
+
+        LOG(f"{self.name}: {hub.battery.voltage()}mV")
+
+        esperado = None
+        if   self.nome == "braço":  esperado = "spike0"
+        elif self.nome == "cabeça": esperado = "spike1"
+        elif self.nome == "rabo":   esperado = "supermini0"
+        else: ASSERT(False, "nome de robô inesperado")
+
+        while name != esperado:
+            hub.speaker.beep(frequency=1024); wait(200)
+        else:
+            hub.light.blink(Color.ORANGE, [100,50,200,100])
+
+    @classmethod
+    def globalizar(self):
+        global TESTE, DEBUG
+        TESTE = self.TESTE
+        DEBUG = self.DEBUG
+        global hub, ble
+        hub = self.hub
+        ble = self.ble
+        global name, nome
+        name = self.name
+        nome = self.nome
+
+class bipes:
+    inicio = lambda: hub.speaker.beep(frequency=500, duration=100)
+    final  = lambda: hub.speaker.beep(frequency=250, duration=250)
+
+    separador  = lambda: hub.speaker.beep(frequency=600, duration=200)
+    calibracao = lambda: hub.speaker.beep(frequency=300, duration=100)
+    cabeca     = lambda: hub.speaker.beep(frequency=600, duration=100)
+
+    falha      = lambda: (hub.speaker.beep(frequency=800, duration=500), wait(200),
+                          hub.speaker.beep(frequency=800, duration=500),)
+
+
+def LOG(*args, print=print, **kwargs):
+    print(f"{nome}:", *args, **kwargs)
+
+def ERRO(*args, bipar=True):
+    LOG("ERRO:", *args);
+    if bipar: bipes.falha()
+
+def ASSERT(cond, texto=None):
+    if DEBUG: assert cond, texto
+    elif not cond:
+        ERRO(f"assert '{texto}' falhou", bipar=DEBUG)
+    return cond
+
